requirements.txt
flask
streamlit
google-generativeai
python-dotenv
psutil
pandas
altair
wmi
pywin32
requests